feat: change db to mysql, applying hashed password and add column role to user

+++ app.py
@@ -1,3 +1,4 @@
+import bcrypt
 from flask import Flask, request, jsonify
 from models.user import User
 from database import db
@@ -11,7 +12,10 @@ from flask_login import (
 
 app = Flask(__name__)
 app.config["SECRET_KEY"] = "key"
-app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///database.db"
+app.config["SQLALCHEMY_DATABASE_URI"] = (
+    "mysql+pymysql://root:admin123@127.0.0.1:3306/flask_crud"
+)
+
 
 login_manager = LoginManager()
 db.init_app(app)
@@ -35,7 +39,7 @@ def login():
     if username and password:
         user = User.query.filter_by(username=username).first()
 
-        if user and user.password == password:
+        if user and bcrypt.checkpw(str.encode(password), str.encode(user.password)):
             login_user(user)
             print(current_user.is_authenticated)
             return jsonify({"message": "Success"})
@@ -57,7 +61,8 @@ def signUp():
     password = data.get("password")
 
     if username and password:
-        user = User(username=username, password=password)
+        hashed_pw = bcrypt.hashpw(str.encode(password), bcrypt.gensalt())
+        user = User(username=username, password=hashed_pw, role="user")
         db.session.add(user)
         db.session.commit()
         return jsonify({"message": "User created successfully"})
@@ -82,6 +87,9 @@ def update_user(id_user):
     data = request.json
     user = User.query.get(id_user)
 
+    if id_user != current_user.id and current_user.role == "user":
+        return jsonify({"message": "Operation not allowed"}), 403
+
     if user and data.get("password"):
         user.password = data.get("password")
         db.session.commit()
@@ -95,6 +103,9 @@ def update_user(id_user):
 @login_required
 def delete_user(id_user):
     user = User.query.get(id_user)
+
+    if current_user.role != "admin":
+        return jsonify({"message": "Operation not allowed"}), 403
 
     if user and id_user != current_user.id:
         db.session.delete(user)
